Fix multiple instance switches
Devices which support multiple endpoints advertise instance 1 as
being the old style non-endpoint connection and instance 2+ typically
correspond to endpoint 1+. So for a dual outlet the mapping would be
inst 1 controls outlet 1 (not associated with an endpoint)
inst 2 controls outlet 1 (associated with endpoint 1)
inst 3 controls outlet 2 (associated with endpoint 2)

With ZW099 Smart Dimmer, it advertises 2 instances, but the second
instance doesn't seem to work properly. So this PR changes things
back to that the 1st outlet is still controlled using instance 1,
and the second and subsequent outlets will be controlled using
instance 3 and above.