Add support for the Aeotec WallMote Quad
Note: that support for the WallMote also requires an updated
config/aeotec/zw130.xml file with the OpenZWave library.

This PR also includes some improvements for battery
powered devices, and come config changes for the Aeotec
Smart Dimmer/Switch so that button presses are detected.

Note: Changing the config on the wall mote (disabling touch,
touch vibration or touch color) requires pressing the button
on the back of the wallmote to enable it to pick up the changes)